pyproject.toml
[project]
name = "deep-research"
version = "0.1.0"
description = "a multi-agent deep research framework that uses smallagents to handle different subtasks."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.7.0",
    "openai>=2.31.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "textual>=8.1.1",
    "textual-dev>=1.8.0",
    "tomli-w>=1.2.0",
]

[tool.setuptools.packages.find]
include = ["research*"]
